internal/tool: move tool integration code into its own package and out of the core code.
