+++ internal/tool/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2021-present The Atlas Authors. All rights reserved.
+// This source code is licensed under the Apache 2.0 license found
+// in the LICENSE file in the root directory of this source tree.
+
+// Package tool contains logic to integrate existing tools like Flyway or Liquibase with the Atlas CLI.
+package tool
